refact: reorganize ui components
