+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/presentation/screens/PokemonSearchScreen.kt
@@ -1,7 +1,12 @@
-package com.github.felipecastilhos.pokedexandroid.features.pokemon
+package com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.screens
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
